requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
dominate==2.3.4
Flask==1.0.2
Flask-Admin==1.5.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysql-connector-python==8.0.13
protobuf==3.6.1
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
qiniu==7.2.2
redis==2.10.6
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.24
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
